+++ views/welcome.php
@@ -124,7 +124,7 @@
 			<form action="<?php echo URL::site("Default/contact") ?>" method="post">
 				<?php
 				$flash = Session::instance()->get("flash", false);
-				if($flash && $flash['source'] == "contact"){
+				if($flash && isset($flash['source']) && "contact" == $flash['source']){
 					echo "<div class='container-narrow'><div class='alert alert-$flash[type]'>$flash[message]</div></div>";
 					Session::instance()->delete("flash");
 				}
